backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.3.6
redis==5.0.4
sqlalchemy==2.0.30
openai-whisper==20231117
ffmpeg-python==0.2.0
requests==2.31.0
python-multipart==0.0.9
pydantic==2.7.1
aiofiles==23.2.1
python-dotenv==1.0.1
